Renamed nativeTransactor to builders (and adjusted the contained files) since we will be having multiple types of builders.
