+++ src/main/scala/co/topl/brambl/builders/TransactionBuilder.scala
@@ -1,9 +1,9 @@
-package co.topl.brambl.nativeTransactor
+package co.topl.brambl.builders
 
 import co.topl.brambl.models.Address
 import co.topl.brambl.models.transaction.IoTransaction
 
-trait Builder {
+trait TransactionBuilder {
   // Construct simple transaction
   def constructTransaction(inputAddress: Address, outputAddress: Address): Option[IoTransaction]
 }
