+++ src/main/scala/co/topl/brambl/wallet/Credentialler.test.sc
@@ -3,7 +3,7 @@
 * How the IoTransaction is created is irrelevant
 * */
 import co.topl.brambl.{Context, QuivrService}
-import co.topl.brambl.nativeTransactor.MockBuilder
+import co.topl.brambl.builders.MockTransactionBuilder
 import co.topl.brambl.wallet.{Credentialler, MockStorage}
 import co.topl.node.box.Locks
 import co.topl.node.{Address, Events, Identifiers}
@@ -38,7 +38,7 @@ def runTest(unprovenTx: IoTransaction, expectedPass: Boolean, createCtx: IoTrans
   }
 }
 
-val builder = MockBuilder(MockStorage)
+val builder = MockTransactionBuilder(MockStorage)
 
 val outputAddress = Address(0, 0, Identifiers.boxLock32(
   Locks.Predicate(List(QuivrService.lockedProposition.get), 1)
